services/fastapi/pyproject.toml
[project]
name = "novitas-fastapi"
version = "0.1.0"
description = "Novitas AI / agent orchestration (FastAPI)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "google-adk>=1.0.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
]
